requirements.txt
pdfplumber>=0.10.0
python-docx>=1.1.0
openpyxl>=3.1.0
reportlab>=4.0.0